2004-03-24  Al Riddoch  <ajr@ecs.soton.ac.uk>
	* app/GlView.cpp, app/GlView.h: Modify the depth frame store to be
	  float. Works better now.

	* app/Holo.cpp: Disable smooth antialiased lines.

	* app/ServerEntities.cpp, app/ServerEntities.h: Clean up code that
	  iterates over the entity tree. Implement select all, and
	  select invert.

+++ app/GlView.h
@@ -30,7 +30,7 @@ class GlView : public Gtk::DrawingArea {
     bool m_redrawRequired;
     bool m_refreshRequired;
     GLubyte * m_frameStore;
-    GLuint * m_depthStore;
+    GLfloat * m_depthStore;
     int m_frameStoreWidth;
     int m_frameStoreHeight;
